Move opencl files to separate folder
