+++ opm/simulators/linalg/bda/WellContributions.cpp
@@ -26,7 +26,7 @@
 #include <opm/simulators/linalg/bda/WellContributions.hpp>
 
 #ifdef HAVE_OPENCL
-#include <opm/simulators/linalg/bda/openclWellContributions.hpp>
+#include <opm/simulators/linalg/bda/opencl/openclWellContributions.hpp>
 #endif
 
 #ifdef HAVE_CUDA
