+++ opm/simulators/linalg/bda/opencl/BILU0.hpp
@@ -25,10 +25,10 @@
 #include <opm/simulators/linalg/bda/BlockedMatrix.hpp>
 #include <opm/simulators/linalg/bda/ILUReorder.hpp>
 
-#include <opm/simulators/linalg/bda/opencl.hpp>
-#include <opm/simulators/linalg/bda/openclKernels.hpp>
+#include <opm/simulators/linalg/bda/opencl/opencl.hpp>
+#include <opm/simulators/linalg/bda/opencl/openclKernels.hpp>
 #include <opm/simulators/linalg/bda/opencl/Preconditioner.hpp>
-#include <opm/simulators/linalg/bda/ChowPatelIlu.hpp>
+#include <opm/simulators/linalg/bda/opencl/ChowPatelIlu.hpp>
 
 
 namespace Opm
