+++ opm/simulators/linalg/bda/opencl/BISAI.cpp
@@ -26,8 +26,8 @@
 #include <dune/common/timer.hh>
 
 #include <opm/simulators/linalg/bda/BdaSolver.hpp>
-#include <opm/simulators/linalg/bda/BILU0.hpp>
-#include <opm/simulators/linalg/bda/BISAI.hpp>
+#include <opm/simulators/linalg/bda/opencl/BILU0.hpp>
+#include <opm/simulators/linalg/bda/opencl/BISAI.hpp>
 #include <opm/simulators/linalg/bda/Reorder.hpp>
 
 namespace Opm
