+++ opm/simulators/linalg/bda/opencl/BISAI.hpp
@@ -22,10 +22,10 @@
 
 #include <mutex>
 
-#include <opm/simulators/linalg/bda/BILU0.hpp>
-#include <opm/simulators/linalg/bda/opencl.hpp>
-#include <opm/simulators/linalg/bda/openclKernels.hpp>
-#include <opm/simulators/linalg/bda/openclSolverBackend.hpp>
+#include <opm/simulators/linalg/bda/opencl/BILU0.hpp>
+#include <opm/simulators/linalg/bda/opencl/opencl.hpp>
+#include <opm/simulators/linalg/bda/opencl/openclKernels.hpp>
+#include <opm/simulators/linalg/bda/opencl/openclSolverBackend.hpp>
 #include <opm/simulators/linalg/bda/opencl/Preconditioner.hpp>
 
 namespace Opm
