+++ opm/simulators/linalg/bda/opencl/CPR.hpp
@@ -30,16 +30,16 @@
 #endif
 #include <dune/istl/umfpack.hh>
 
-#include <opm/simulators/linalg/bda/BILU0.hpp>
-#include <opm/simulators/linalg/bda/opencl.hpp>
+#include <opm/simulators/linalg/bda/opencl/BILU0.hpp>
+#include <opm/simulators/linalg/bda/opencl/opencl.hpp>
 #include <opm/simulators/linalg/bda/Matrix.hpp>
-#include <opm/simulators/linalg/bda/OpenclMatrix.hpp>
+#include <opm/simulators/linalg/bda/opencl/OpenclMatrix.hpp>
 #include <opm/simulators/linalg/bda/ILUReorder.hpp>
 #include <opm/simulators/linalg/bda/opencl/Preconditioner.hpp>
 
-#include <opm/simulators/linalg/bda/openclKernels.hpp>
-#include <opm/simulators/linalg/bda/ChowPatelIlu.hpp>
-#include <opm/simulators/linalg/bda/openclSolverBackend.hpp>
+#include <opm/simulators/linalg/bda/opencl/openclKernels.hpp>
+#include <opm/simulators/linalg/bda/opencl/ChowPatelIlu.hpp>
+#include <opm/simulators/linalg/bda/opencl/openclSolverBackend.hpp>
 
 namespace Opm
 {
