+++ opm/simulators/linalg/bda/opencl/opencl.cpp
@@ -18,7 +18,7 @@
 */
 
 #include <config.h>
-#include <opm/simulators/linalg/bda/opencl.hpp>
+#include <opm/simulators/linalg/bda/opencl/opencl.hpp>
 #include <string>
 
 namespace Opm
