+++ tests/test_csrToCscOffsetMap.cpp
@@ -10,7 +10,7 @@
 #include <dune/istl/bcrsmatrix.hh>
 #include <dune/istl/matrixmarket.hh>
 
-#include <opm/simulators/linalg/bda/BISAI.hpp>
+#include <opm/simulators/linalg/bda/opencl/BISAI.hpp>
 
 BOOST_AUTO_TEST_CASE(testcsrtocscoffsetmap){
